# Speed up `complete_levels()`