--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -1,6 +1,6 @@
 Package: marginaleffects
 Title: Predictions, Comparisons, Slopes, Marginal Means, and Hypothesis Tests
-Version: 0.9.0.9045
+Version: 0.9.0.9046
 Authors@R: 
     c(person(given = "Vincent",
              family = "Arel-Bundock",