Example coming from the [`marginal_means()` vignette](https://vincentarelbundock.github.io/marginaleffects/dev/articles/marginalmeans.html#case-study-multinomial-logit) (just reduced the number in `replicate()`).

Before:
``` r
library(nnet)
library(marginaleffects)

set.seed(1839)
n <- 1200
x <- factor(sample(letters[1:3], n, TRUE))
y <- vector(length = n)
y[x == "a"] <- sample(letters[4:6], sum(x == "a"), TRUE)
y[x == "b"] <- sample(letters[4:6], sum(x == "b"), TRUE, c(1 / 4, 2 / 4, 1 / 4))
y[x == "c"] <- sample(letters[4:6], sum(x == "c"), TRUE, c(1 / 5, 3 / 5, 2 / 5))

dat <- data.frame(x = x, y = factor(y))
tmp <- as.data.frame(replicate(9, factor(sample(letters[7:9], n, TRUE))))
dat <- cbind(dat, tmp)
void <- capture.output({
  mod <- multinom(y ~ ., dat)
})

bench::mark(
  expr = marginal_means(mod,
                        type = "probs",
                        variables = c("x", "V1"),
                        variables_grid = paste0("V", 2:3)),
  iterations = 10
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 expr          25.8s    26.1s    0.0372    13.9GB     7.81
```

After:
``` r
bench::mark(
  expr = marginal_means(mod,
                        type = "probs",
                        variables = c("x", "V1"),
                        variables_grid = paste0("V", 2:3)),
  iterations = 10
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 expr          17.7s    18.6s    0.0533    13.8GB     13.2
```